+++ arctic_platform/client/__init__.py
@@ -12,19 +12,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from arctic_platform.client.client import ArcticClient
-from arctic_platform.client.client import ArcticRLClient
-from arctic_platform.client.client import ArcticSFTClient
-from arctic_platform.client.client import SyncArcticRLClient
-from arctic_platform.client.client import create_arctic_rl_client
-from arctic_platform.client.client import create_arctic_sft_client
-from arctic_platform.client.client import make_transport
+from arctic_platform.client.base import ArcticClient
+from arctic_platform.client.base import AsyncArcticClient
+from arctic_platform.client.base import SyncArcticClient
+from arctic_platform.client.base import make_transport
 from arctic_platform.client.config import ArcticClientConfig
 from arctic_platform.client.config import ArcticRLClientConfig
 from arctic_platform.client.config import CortexConfig
 from arctic_platform.client.config import OnPremConfig
 from arctic_platform.client.config import SamplingConfig
 from arctic_platform.client.config import TrainingConfig
+from arctic_platform.client.rl import ArcticRLClient
+from arctic_platform.client.rl import SyncArcticRLClient
+from arctic_platform.client.rl import create_arctic_rl_client
+from arctic_platform.client.sft import ArcticSFTClient
+from arctic_platform.client.sft import create_arctic_sft_client
 from arctic_platform.client.transport import OPS
 from arctic_platform.client.transport import JobHandles
 from arctic_platform.client.transport import Request
@@ -38,13 +40,15 @@ __all__ = [
     "ArcticRLClient",
     "ArcticRLClientConfig",
     "ArcticSFTClient",
+    "AsyncArcticClient",
     "CortexConfig",
     "JobHandles",
     "OnPremConfig",
     "SamplingConfig",
+    "SyncArcticClient",
+    "SyncArcticRLClient",
     "TrainingConfig",
     "Request",
-    "SyncArcticRLClient",
     "Transport",
     "create_arctic_rl_client",
     "create_arctic_sft_client",
