--- arctic_platform/sft/client.py
@@ -1,20 +1,0 @@
-# Copyright 2025 Snowflake Inc.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Back-compat shim — the SFT client now lives in ``arctic_platform.client.client``."""
-
-from arctic_platform.client.client import ArcticSFTClient
-from arctic_platform.client.client import create_arctic_sft_client
-
-__all__ = ["ArcticSFTClient", "create_arctic_sft_client"]
